+++ client/src/App.js
@@ -21,9 +21,9 @@ import Order from "./pages/Order";
 import Contact from "./pages/Contact";
 import Subscribe from "./pages/Subscribe";
 import Profile from "./pages/Profile";
+// import Cart from "./components/Cart/Cart";
 
 // import OrderHistory from "./pages/OrderHistory";
-// import Cart from "./pages/Cart";
 const httpLink = createHttpLink({
   uri: "/graphql",
 });
@@ -62,6 +62,7 @@ export default function App() {
                 <Route exact path="/contact" element={<Contact />} />
                 <Route exact path="/subscribe" element={<Subscribe />} />
                 <Route exact path="/profile" element={<Profile />} />
+                {/* <Route exact path="/cart" element={<Cart />} /> */}
               </Routes>
               <Footer />
             </div>
