+++ client/src/components/Products/ProductList.js
@@ -1,10 +1,15 @@
 import React, { useEffect } from "react";
 // import ProductItem from "../Products/ProductItem";
-// import { useStoreContext } from "../../utils/GlobalState";
+import { useStoreContext } from "../../utils/GlobalState";
 // import { UPDATE_PRODUCTS } from "../../utils/actions";
+import { ADD_TO_CART, UPDATE_CART_QUANTITY } from "../../utils/actions";
+
 import { useQuery } from "@apollo/client";
 import { QUERY_ALL_PRODUCTS } from "../../utils/queries";
 import Image from "../../assets/images/scoop.png";
+import { Link } from "react-router-dom";
+import cartImg from "../../assets/images/add-cart.png";
+import { idbPromise } from "../../utils/helpers";
 
 const styles = {
   scoop: {
@@ -20,14 +25,46 @@ const styles = {
     fontFamily: "Syncopate",
     fontDisplay: "sans-serif",
     textAlign: "center",
+    textDecoration: "none",
+  },
+  img: {
+    height: "2.5rem",
+    width: "2.5rem",
+    invert: 1,
   },
 };
-function ProductList() {
+
+function ProductList(item) {
+  const [state, dispatch] = useStoreContext();
+
+  const { flavor, _id, price, size } = item;
+  const { cart } = state;
+
   const { loading, data } = useQuery(QUERY_ALL_PRODUCTS);
 
   const products = data?.products || [];
 
   console.log(products);
+  const addToCart = () => {
+    const itemInCart = cart.find((cartItem) => cartItem._id === _id);
+    if (itemInCart) {
+      dispatch({
+        type: UPDATE_CART_QUANTITY,
+        _id: _id,
+        purchaseQuantity: parseInt(itemInCart.purchaseQuantity) + 1,
+      });
+      idbPromise("cart", "put", {
+        ...itemInCart,
+        purchaseQuantity: parseInt(itemInCart.purchaseQuantity) + 1,
+      });
+    } else {
+      dispatch({
+        type: ADD_TO_CART,
+        product: { ...item, purchaseQuantity: 1 },
+      });
+      idbPromise("cart", "put", { ...item, purchaseQuantity: 1 });
+    }
+  };
   return (
     <div className="my-2">
       <h2></h2>
@@ -40,11 +77,11 @@ function ProductList() {
         </div>
       ) : (
         products.map((product) => (
-          <div className="row p-5 justify-content-around">
-            <div className="d-flex flex-column my-5 col-xl-3 col-lg-4 col-md-5 col-sm-6">
+          <div className="row p-2 justify-content-around">
+            <div className="d-flex flex-column my-3 col-xl-3 col-lg-4 col-md-5 col-sm-6">
               <div
                 style={styles.card}
-                className="card align-items-center text-center"
+                className="card align-items-center text-center p-3"
               >
                 <img
                   style={styles.scoop}
@@ -52,32 +89,25 @@ function ProductList() {
                   className="card-img-top"
                 />
                 <div className="card-body">
-                  <h5 className="card-title" style={styles.flavorites}>
-                    {product.flavor}
-                  </h5>
+                  <Link to={`/products/${product._id}`}>
+                    <h5 className="card-title" style={styles.flavorites}>
+                      {product.flavor}
+                    </h5>
+                  </Link>
+                  <span>${product.price}</span>
                 </div>
+
+                <button
+                  onClick={addToCart}
+                  style={styles.img}
+                  src={cartImg}
+                  className="btn bg-transparent"
+                ></button>
               </div>
             </div>
           </div>
         ))
       )}
-
-      {/* {state.products.length ? (
-        <div className="flex-row">
-          {filterProducts().map((product) => (
-            <ProductItem
-              // key={product._id}
-              _id={product._id}
-              // image={product.image}
-              flavor={product.flavor}
-              price={product.price}
-              size={product.size}
-            />
-          ))}
-        </div>
-      ) : (
-        <h3>You haven't added any products yet!</h3>
-      )} */}
     </div>
   );
 }
